% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{abundant_isotopes}
\alias{abundant_isotopes}
\title{abundant_isotopes}
\format{A data frame with 97 rows and 5 variables:
\describe{
  \item{\code{element}}{character ##}
  \item{\code{isotope}}{character ##}
  \item{\code{mass}}{double ##}
  \item{\code{abundance}}{double ##}
  \item{\code{ratioC}}{integer ##}
}}
\usage{
abundant_isotopes
}
\description{
This table include the abundant isotope used to calculate exact mass for given formula
}
\details{
DETAILS
}
\keyword{datasets}
