% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elem_num_query.R
\name{elem_num_query}
\alias{elem_num_query}
\title{elem_num_query}
\usage{
elem_num_query(formula, elem_query)
}
\arguments{
\item{formula}{for example "C2H4O2"}

\item{elem_query}{it refers to element in query}
}
\value{
It returns the number of element in formula
}
\description{
elem_num_query
}
\examples{
elem_num_query("C2H4O2","C")
}
