% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{elem_table}
\alias{elem_table}
\title{elem_table}
\format{A data frame with 30 rows and 5 variables:
\describe{
  \item{\code{element}}{character COLUMN_DESCRIPTION}
  \item{\code{isotope}}{character COLUMN_DESCRIPTION}
  \item{\code{mass}}{double COLUMN_DESCRIPTION}
  \item{\code{abundance}}{double COLUMN_DESCRIPTION}
  \item{\code{unsaturation}}{double COLUMN_DESCRIPTION}
}}
\usage{
elem_table
}
\description{
elem_table
}
\details{
DETAILS
}
\keyword{datasets}
