% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula_mz.R
\name{formula_mz}
\alias{formula_mz}
\title{formula_mz}
\usage{
formula_mz(abundant_isotopes, formula = "C2H4O1", charge = 0)
}
\arguments{
\item{abundant_isotopes}{a table can be imported}

\item{formula}{e.g. "C2H4O1"}

\item{charge}{0: neutral, 1: positive charge, -1: negative charge}
}
\value{
the exact mass of the input formula
}
\description{
formula_mz
}
\examples{
formula_mz(abundant_isotopes, formula = "C2H4O1", charge = 0)


}
