% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula_rdbe.R
\name{formula_rdbe}
\alias{formula_rdbe}
\title{formula_rdbe}
\usage{
formula_rdbe(formula = "C2H4O1", elem_table = lc8::elem_table)
}
\arguments{
\item{formula}{e.g. "C2H4O1"}

\item{elem_table}{a table records unsaturation}
}
\value{
the ring and double bond number
}
\description{
formula_rdbe
}
\examples{
formula_rdbe(formula = "C2H4O1", elem_table = lc8::elem_table)
}
