% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{full_isotopes}
\alias{full_isotopes}
\title{full_isotopes}
\format{A data frame with 308 rows and 5 variables:
\describe{
  \item{\code{element}}{character COLUMN_DESCRIPTION}
  \item{\code{isotope}}{character COLUMN_DESCRIPTION}
  \item{\code{mass}}{double COLUMN_DESCRIPTION}
  \item{\code{abundance}}{double COLUMN_DESCRIPTION}
  \item{\code{ratioC}}{integer COLUMN_DESCRIPTION}
}}
\usage{
full_isotopes
}
\description{
This table include the abundant isotope used to calculate exact mass for given formula
}
\details{
DETAILS
}
\keyword{datasets}
