% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_break_formula.R
\name{my_break_formula}
\alias{my_break_formula}
\title{my_break_formula}
\usage{
my_break_formula(formula = "[13]C1C-1H4N2")
}
\arguments{
\item{formula}{any chemical formula}
}
\value{
a list containing elem and count
}
\description{
my_break_formula
}
\examples{
my_break_formula("[13]C1C-1H4N2")


}
