% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_calculate_formula.R
\name{my_calculate_formula}
\alias{my_calculate_formula}
\title{my_calculate_formula}
\usage{
my_calculate_formula(formula1, formula2, sign = 1, Is_valid = TRUE)
}
\arguments{
\item{formula1}{formula1}

\item{formula2}{formula1}

\item{sign}{1 for merge two formula, -1 for subtract second from first formula}

\item{Is_valid}{if set True, only return formula if the element count is > 0}
}
\value{
the merge or subtract formula
}
\description{
my_calculate_formula
}
\examples{
my_calculate_formula(formula1 = "C2H4O1S2P1",formula2 = "N1H1O-1",sign = 1,Is_valid = TRUE)
}
