% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_subfrom.R
\name{My_subfrom}
\alias{My_subfrom}
\title{My_subfrom}
\usage{
My_subfrom(formula1, formula2)
}
\arguments{
\item{formula1}{enter a chemical formula, e.g. "C2H4O2"}

\item{formula2}{enter a second formula, e.g. "C1H2"}
}
\value{
Subtract formula1 to formula2
}
\description{
My_subfrom
}
\examples{
My_subfrom("C2H4O2","C1H2")


}
