% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mz_formula.R
\name{mz_formula}
\alias{mz_formula}
\title{mz_formula}
\usage{
mz_formula(Accurate_mass = 148.0604, charge = 1, ppm = 5,
  C_range = 0:99, H_range = 0:100, O_range = 0:20, N_range = 0:12,
  Cl_range = 0:0, P_range = 0:3, S_range = 0:3, Na_range = 0:0,
  K_range = 0:0, F_range = 0:0, Br_range = 0:0, I_range = 0:0,
  Si_range = 0:0, B_range = 0:0, Ca_range = 0:0, Cu_range = 0:0,
  Ni_range = 0:0, N_rule = T, db_min = 0, db_max = 99,
  metal_ion = 0:3)
}
\arguments{
\item{Accurate_mass}{accurate mass from ms data}

\item{charge}{1 for positive, 0 for neutral mass, -1 for negative}

\item{ppm}{output result tolerance window}

\item{C_range}{carbon number range}

\item{H_range}{hydrogen number range}

\item{O_range}{oxygen number range}

\item{N_range}{nitrogen number range}

\item{Cl_range}{chlorine number range}

\item{P_range}{phosphorus number range}

\item{S_range}{sulfur number range}

\item{Na_range}{sodium number range}

\item{K_range}{potassium number range}

\item{F_range}{fluorine number range}

\item{Br_range}{bromine number range}

\item{I_range}{iodine number range}

\item{Si_range}{silicon number range}

\item{B_range}{boron number range}

\item{Ca_range}{Ca number range}

\item{Cu_range}{Cu number range}

\item{Ni_range}{Ni number range}

\item{N_rule}{nitrogen rule}

\item{db_min}{min of double bond and ring number}

\item{db_max}{max of double bond and ring number}

\item{metal_ion}{total number of metal ion}
}
\value{
a dataframe recording formula, mass difference and db_r number that fall within ppm of input mass
}
\description{
mz_formula
}
\examples{
mz_formula(148.0604, 1, 5)

}
