% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_mergefrom.R
\name{My_mergefrom}
\alias{My_mergefrom}
\title{My_mergefrom}
\usage{
My_mergefrom(formula1, formula2)
}
\arguments{
\item{formula1}{enter a chemical formula, e.g. "C2H4O2"}

\item{formula2}{enter a second formula, e.g. "C1H3Cl2"}
}
\value{
the merge formula of two entered formula
}
\description{
My_mergefrom
}
\examples{
My_mergefrom("C2H4O2","C1H3Cl2")


}
